refactor: implement review recomendations
