+++ packages/client/src/game/services/AudioManager/AudioManager.test.ts
@@ -11,10 +11,10 @@ function mockEntity(rect: Rect) {
 
 describe('game/services/AudioManager', () => {
   const mockAudioContext = {} as AudioContext;
+  const mockGame = { resources: { audioContext: mockAudioContext } } as Game;
 
   it('should play sounds', () => {
-    const audioManager = new AudioManager({} as Game);
-    audioManager.context = mockAudioContext;
+    const audioManager = new AudioManager(mockGame);
 
     audioManager.playSound = jest.fn();
 
@@ -33,8 +33,7 @@ describe('game/services/AudioManager', () => {
   });
 
   it('should play pause sound', () => {
-    const audioManager = new AudioManager({} as Game);
-    audioManager.context = mockAudioContext;
+    const audioManager = new AudioManager(mockGame);
 
     audioManager.playSound = jest.fn();
 
@@ -44,8 +43,7 @@ describe('game/services/AudioManager', () => {
   });
 
   it('should not play sounds while paused', () => {
-    const audioManager = new AudioManager({} as Game);
-    audioManager.context = mockAudioContext;
+    const audioManager = new AudioManager(mockGame);
 
     const entity = mockEntity({ posX: 2, posY: 2, width: 2, height: 2 });
