+++ packages/client/src/game/services/AudioManager/typings.ts
@@ -1,4 +1,4 @@
-import { type SoundPathList } from '../Resources/data';
+import { SoundPathList } from '../Resources/data';
 
 export type ActivatedSounds = Record<keyof typeof SoundPathList, ActiveSound>;
 
@@ -9,3 +9,7 @@ type ActiveSound = {
   resumeFrom: number;
   isEnded?: boolean;
 };
+
+export function isSoundNameInSoundPathList(soundName: string): soundName is keyof typeof SoundPathList {
+  return soundName in SoundPathList;
+}
