+++ packages/client/src/tests/setup.ts
@@ -33,3 +33,21 @@ jest.mock('react-router-dom', () => {
     ScrollRestoration: () => null,
   };
 });
+
+const AudioContext = jest.fn(() => ({
+  createBufferSource: jest.fn(() => ({
+    buffer: null,
+    connect: jest.fn(),
+    disconnect: jest.fn(),
+    start: jest.fn(),
+    stop: jest.fn(),
+  })),
+  createBuffer: jest.fn(),
+  decodeAudioData: jest.fn(),
+  currentTime: 0,
+  resume: jest.fn(),
+  suspend: jest.fn(),
+}));
+
+//@ts-expect-error - это мок, у него нет всех свойств реального AudioContext
+window.AudioContext = AudioContext;
